# backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: ShelterFlex Backend API
  description: API for ShelterFlex backend service
  version: 0.1.0
  contact:
    name: ShelterFlex Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.shelterflex.io
    description: Production server
paths:
...
  /api/admin/fraud/slashing/submit-evidence:
    post:
      summary: Submit evidence to slashing module
      description: Queues a submit_evidence operation to the slashing module contract via the outbox pattern.
      operationId: submitSlashingEvidence
      tags:
        - Admin
      security:
        - adminSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - submitter
                - commitment
                - actor
                - offence
              properties:
                submitter:
                  type: string
                  description: Stellar address of the evidence submitter
                  example: "GABC..."
                commitment:
                  type: string
                  description: Commitment hash (hex string)
                  example: "0x123abc..."
                actor:
                  type: string
                  description: Stellar address of the actor being accused
                  example: "GXYZ..."
                offence:
                  type: string
                  description: Type of offence
                  example: "double_signing"
      responses:
        '201':
          description: Evidence submission queued
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - outboxId
                properties:
                  success:
                    type: boolean
                    example: true
                  outboxId:
                    type: string
                    description: Outbox item ID for tracking
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Invalid admin secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: FORBIDDEN
                  message: Invalid admin secret
        default:
          $ref: '#/components/responses/Error'
  /api/admin/fraud/slashing/reveal-evidence:
    post:
      summary: Reveal evidence in slashing module
      description: Queues a reveal_evidence operation to the slashing module contract via the outbox pattern.
      operationId: revealSlashingEvidence
      tags:
        - Admin
      security:
        - adminSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - submitter
                - slashId
                - evidence
                - salt
              properties:
                submitter:
                  type: string
                  description: Stellar address of the evidence submitter
                  example: "GABC..."
                slashId:
                  type: integer
                  description: Slash ID from the original submission
                  example: 1
                evidence:
                  type: string
                  description: Evidence data (hex string)
                  example: "0x456def..."
                salt:
                  type: string
                  description: Salt used for commitment (hex string)
                  example: "0x789ghi..."
      responses:
        '201':
          description: Evidence reveal queued
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - outboxId
                properties:
                  success:
                    type: boolean
                    example: true
                  outboxId:
                    type: string
                    description: Outbox item ID for tracking
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Invalid admin secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: FORBIDDEN
                  message: Invalid admin secret
        default:
          $ref: '#/components/responses/Error'
  /api/admin/fraud/slashing/propose-slash:
    post:
      summary: Propose a slash in slashing module
      description: Queues a propose_slash operation to the slashing module contract via the outbox pattern.
      operationId: proposeSlash
      tags:
        - Admin
      security:
        - adminSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - submitter
                - actor
                - penaltyBps
              properties:
                submitter:
                  type: string
                  description: Stellar address of the submitter
                  example: "GABC..."
                actor:
                  type: string
                  description: Stellar address of the actor to be slashed
                  example: "GXYZ..."
                penaltyBps:
                  type: integer
                  description: Penalty in basis points (0-10000, where 10000 = 100%)
                  minimum: 0
                  maximum: 10000
                  example: 5000
      responses:
        '201':
          description: Slash proposal queued
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - outboxId
                properties:
                  success:
                    type: boolean
                    example: true
                  outboxId:
                    type: string
                    description: Outbox item ID for tracking
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Invalid admin secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: FORBIDDEN
                  message: Invalid admin secret
        default:
          $ref: '#/components/responses/Error'
  /api/admin/fraud/slashing/finalize-slash:
    post:
      summary: Finalize a slash in slashing module
      description: Queues a finalize_slash operation to the slashing module contract via the outbox pattern.
      operationId: finalizeSlash
      tags:
        - Admin
      security:
        - adminSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - caller
                - slashId
              properties:
                caller:
                  type: string
                  description: Stellar address of the caller
                  example: "GABC..."
                slashId:
                  type: integer
                  description: Slash ID to finalize
                  example: 1
      responses:
        '201':
          description: Slash finalization queued
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - outboxId
                properties:
                  success:
                    type: boolean
                    example: true
                  outboxId:
                    type: string
                    description: Outbox item ID for tracking
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Invalid admin secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: FORBIDDEN
                  message: Invalid admin secret
        default:
          $ref: '#/components/responses/Error'
  /api/admin/fraud/slashing/cancel-slash:
    post:
      summary: Cancel a slash in slashing module
      description: Queues a cancel_slash operation to the slashing module contract via the outbox pattern.
      operationId: cancelSlash
      tags:
        - Admin
      security:
        - adminSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - admin
                - slashId
              properties:
                admin:
                  type: string
                  description: Stellar address of the admin
                  example: "GABC..."
                slashId:
                  type: integer
                  description: Slash ID to cancel
                  example: 1
      responses:
        '201':
          description: Slash cancellation queued
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - outboxId
                properties:
                  success:
                    type: boolean
                    example: true
                  outboxId:
                    type: string
                    description: Outbox item ID for tracking
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Invalid admin secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: FORBIDDEN
                  message: Invalid admin secret
        default:
          $ref: '#/components/responses/Error'
  /api/admin/fraud/bond/deposit:
    post:
      summary: Deposit bond to bond collateral contract
      description: Queues a deposit_bond operation to the bond collateral contract via the outbox pattern.
      operationId: depositBond
      tags:
        - Admin
      security:
        - adminSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inspector
                - amount
              properties:
                inspector:
                  type: string
                  description: Stellar address of the inspector
                  example: "GABC..."
                amount:
                  type: string
                  description: Bond amount (as string to preserve precision)
                  example: "1000000"
      responses:
        '201':
          description: Bond deposit queued
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - outboxId
                properties:
                  success:
                    type: boolean
                    example: true
                  outboxId:
                    type: string
                    description: Outbox item ID for tracking
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Invalid admin secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: FORBIDDEN
                  message: Invalid admin secret
        default:
          $ref: '#/components/responses/Error'
  /api/admin/fraud/bond/withdraw:
    post:
      summary: Withdraw bond from bond collateral contract
      description: Queues a withdraw_bond operation to the bond collateral contract via the outbox pattern.
      operationId: withdrawBond
      tags:
        - Admin
      security:
        - adminSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inspector
                - amount
              properties:
                inspector:
                  type: string
                  description: Stellar address of the inspector
                  example: "GABC..."
                amount:
                  type: string
                  description: Bond amount to withdraw (as string to preserve precision)
                  example: "1000000"
      responses:
        '201':
          description: Bond withdrawal queued
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - outboxId
                properties:
                  success:
                    type: boolean
                    example: true
                  outboxId:
                    type: string
                    description: Outbox item ID for tracking
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Invalid admin secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: FORBIDDEN
                  message: Invalid admin secret
        default:
          $ref: '#/components/responses/Error'
...
components:
...
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - TOO_MANY_REQUESTS
                - NOT_FOUND
                - CONFLICT
                - LISTING_ALREADY_RENTED
                - SOROBAN_ERROR
                - INTERNAL_ERROR
              example: NOT_FOUND
            message:
              type: string
              example: "Resource not found"
            details:
              type: object
              description: Optional structured context for the error
              additionalProperties: true
...
  responses:
    Error:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Request validation failed"
              details:
                fieldErrors:
                  - field: "email"
                    message: "Invalid email address"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: >-
        Ordinary authenticated-session bearer token (authenticateToken middleware only —
        no elevated role/permission required).
    bearerAdminAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: >-
        Same bearer token as bearerAuth, but the route additionally requires the caller's
        session to carry an admin-tier RBAC permission (requirePermission(resource, action)
        in middleware/rbac.ts, e.g. payouts:trigger) or an explicit admin/super_admin role
        check. Kept as a distinct scheme name so admin-gated operations are visibly
        distinguishable from ordinary authenticated ones in this document.
    adminSecret:
      type: apiKey
      in: header
      name: x-admin-secret
      description: >-
        Shared operator secret checked against MANUAL_ADMIN_SECRET. Used by the
        job-scheduler admin routes, which are reachable without a user session.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: >-
        Provider-specific HMAC signature header verified server-side instead of a bearer
        token (e.g. x-paystack-signature for Paystack, verif-hash for Flutterwave). The
        header name varies by payment rail; see the payments webhook handler.
...